+++ packages/config/src/mutations/update.js
@@ -3,11 +3,15 @@
 const { writeFile, unlink } = require('fs')
 const { promisify } = require('util')
 
+const cpFile = require('cp-file')
+const makeDir = require('make-dir')
 const pathExists = require('path-exists')
 
 const { ensureConfigPriority } = require('../context')
 const { mergeConfigs } = require('../merge')
 const { parseOptionalConfig } = require('../parse')
+const { addConfigRedirects } = require('../redirects')
+const { simplifyConfig } = require('../simplify')
 const { serializeToml } = require('../utils/toml')
 
 const { applyMutations } = require('./apply')
@@ -17,15 +21,18 @@ const pUnlink = promisify(unlink)
 
 // Persist configuration changes to `netlify.toml`.
 // If `netlify.toml` does not exist, creates it. Otherwise, merges the changes.
-const updateConfig = async function (configMutations, { configPath, redirectsPath, context, branch }) {
+const updateConfig = async function (configMutations, { buildDir, configPath, redirectsPath, context, branch }) {
   if (configMutations.length === 0) {
     return
   }
 
   const inlineConfig = applyMutations({}, configMutations)
-  const normalizedInlineConfig = ensureConfigPriority(inlineConfig, context, branch)
+  const simplifiedConfig = simplifyConfig(inlineConfig)
+  const normalizedInlineConfig = ensureConfigPriority(simplifiedConfig, context, branch)
   const updatedConfig = await mergeWithConfig(normalizedInlineConfig, configPath)
-  await Promise.all([saveConfig(configPath, updatedConfig), deleteRedirectsFile(redirectsPath, normalizedInlineConfig)])
+  const finalConfig = await addConfigRedirects(updatedConfig, redirectsPath)
+  await backupConfig({ buildDir, configPath, redirectsPath })
+  await Promise.all([saveConfig(configPath, finalConfig), deleteRedirectsFile(redirectsPath, normalizedInlineConfig)])
 }
 
 // If `netlify.toml` exists, deeply merges the configuration changes.
@@ -36,8 +43,8 @@ const mergeWithConfig = async function (normalizedInlineConfig, configPath) {
 }
 
 // Serialize the changes to `netlify.toml`
-const saveConfig = async function (configPath, updatedConfig) {
-  const serializedConfig = serializeToml(updatedConfig)
+const saveConfig = async function (configPath, finalConfig) {
+  const serializedConfig = serializeToml(finalConfig)
   await pWriteFile(configPath, serializedConfig)
 }
 
@@ -55,4 +62,47 @@ const deleteRedirectsFile = async function (redirectsPath, normalizedInlineConfi
   await pUnlink(redirectsPath)
 }
 
-module.exports = { updateConfig }
+// Modifications to `netlify.toml` and `_redirects` are only meant for the
+// deploy API call. After it's been performed, we restore their former state.
+// We do this by backing them up inside some sibling directory.
+const backupConfig = async function ({ buildDir, configPath, redirectsPath }) {
+  const tempDir = getTempDir(buildDir)
+  await makeDir(tempDir)
+  await Promise.all([
+    copyIfExists(configPath, `${tempDir}/netlify.toml`),
+    copyIfExists(redirectsPath, `${tempDir}/_redirects`),
+  ])
+}
+
+const restoreConfig = async function ({ buildDir, configPath, redirectsPath }) {
+  const tempDir = getTempDir(buildDir)
+  await Promise.all([
+    copyOrDelete(`${tempDir}/netlify.toml`, configPath),
+    copyOrDelete(`${tempDir}/_redirects`, redirectsPath),
+  ])
+}
+
+const getTempDir = function (buildDir) {
+  return `${buildDir}/.netlify/deploy`
+}
+
+const copyIfExists = async function (src, dest) {
+  if (!(await pathExists(src))) {
+    return
+  }
+
+  await cpFile(src, dest)
+}
+
+const copyOrDelete = async function (src, dest) {
+  if (await pathExists(src)) {
+    await cpFile(src, dest)
+    return
+  }
+
+  if (await pathExists(dest)) {
+    await pUnlink(dest)
+  }
+}
+
+module.exports = { updateConfig, restoreConfig }